Varie Modifiche Helper

+++ NuovoCinemaParadiso/Dtos/DtoAbbonamento.cs
@@ -5,6 +5,6 @@ public class DtoAbbonamento
     public string? Id {get; set;}
     public string Nome {get; set;} = string.Empty;
     public int Durata{get; set;}
-    public decimal Prezzo {get; set;}
+    public int Prezzo {get; set;}
     public int Sconto {get; set;}
 }
